New implementation of interrupts, see complete description below
This is a lightweight and working implementation of interrupts inside klee.
When an interrupt arrives, we do not fork another state, we simply call the
corresponding handler.
This way the handler sees the correct state of variables.
Moreover, this allows to change state while serving an interrupt and to change
state inside an interrupt e.g. if there is a symbolic branch.

All demos up to know are working.

LIMITATIONS:
1. the mapping between interrupt id and handler name is still static (json file)
2. nested interrupts are still not allowed